backend/requirements.txt
fastapi[all]
supabase
openai
pydantic
pydantic-settings
python-dotenv
pyjwt
